Fix docs generation (#26)
* Replace whitelist_extenrals

It was deprecated in tox 3.18.0 in favor of allowlist_externals, and
is no longer available in tox > 4.0 .

* Replace remaining usage of distutils

distutils was deprecated in Python 3.10, and was removed from core
Python in Python 3.12 .

It was removed from code by 25360d5a2c1722ca56d9cbdceba09964f466a00e
but was kept in documentation config.

* Drop unicode prefix

In python 3 all strings are unicode.